moving common SAS code to separate file
